+++ src/utils/utils.ts
@@ -1,4 +1,4 @@
-import { Transducer } from "../data/data";
+import { Transducer, FormState } from "../models/model";
 
 export const formatDate = (date: string): string => {
   const shortDate = new Date(date.split('T', 1).toString());
@@ -39,4 +39,23 @@ export const setExpirationDate = (hoursTillExpiration: number) => {
   }
 
   return new Date(new Date().getTime() + 1000 * 60 * 60 * hours);
+};
+
+// Extract already created transducer's data to populate edit form
+export const transducerFormValues = (transducer: Transducer, condition: string): FormState => {
+  const newFormData: FormState = {
+    name: transducer.name,
+    location: transducer.location,
+    department: transducer.department,
+    room: transducer.room,
+    type: transducer.transducerType,
+    serial: transducer.serialNumber,
+    internal: transducer.internalIdentifier,
+    control: transducer.controlNumber,
+    condition,
+    service: transducer.outOfService,
+    notes: '',
+  };
+
+  return newFormData;
 };
