+++ tests/data/testData.ts
@@ -1,29 +1,4 @@
-export type Condition = 'New' | 'Working' | 'Refurbished' | 'Loaner' | 'Broken (Out of Service)';
-export type Location = 'CMC' | 'MIDTOWN' | 'RISMAN' | 'CROCKER' | 'STREETSBORO' | 'BETTY THE BUS';
-export type Department = 'MFM' | 'L&D' | 'IVF';
-export type TransducerType = 'TA' | 'TV';
-
-export interface TransducerCondition {
-  id: string;
-  condition: Condition;
-  conditionChangedDate: string;
-  note: string;
-}
-
-export interface Transducer {
-  id: string;
-  name: string;
-  location: Location;
-  department: Department;
-  transducerType: TransducerType;
-  room: string;
-  serialNumber: string;
-  internalIdentifier: string;
-  controlNumber: string;
-  dateReceived: string;
-  outOfService: boolean;
-  currentCondition: TransducerCondition[];
-}
+import { Transducer } from "../../src/models/model";
 
 export const TRANSDUCERS: Transducer[] = [
   {
